Only test this pre-1.10.
Django 1.10 drastically changed how lookups worked, and as a result, all
of the lookups are registered by default. Since we only want the default
behaviour of the appropriate ones anyway, changing to a custom set would
be more effort than it's worth, for a feature that introduces
relatively little safety.