Better error for missing but required field
Currently if you fail to provide an enum value for an enumfield
which doesn't have a default you get an `AttributeError` which
isn't very useful as it doesn't tell you which field is missing.

Change the way that we prepare values for the database such that
we instead rely on Django's handling of missing fields.

This means that we now also treat empty strings as nulls, which
is probably fine.