Fix import for migration deconstruction
This field moved its "canonical" location a long time ago (tests pass on supported versions of Django), and while the `.fields` location works, it results in unstable migrations when an EnumField is compared to previous migrations, for a compound field comprised of an EnumField and something else. A motivating use-case for this would be a `MoneyField` that combines a `DecimalField` and an `EnumField`.